Adjust docs and benchmark tools

+++ bin/bench/index.php
@@ -2,17 +2,20 @@
 
 $config = require_once __DIR__ . '/../config.php';
 
-$options = getopt('l::du', ['limit::', 'debug', 'update']);
+$options = getopt('l::du', ['limit::', 'debug', 'update', 'no-delete']);
 $limit = intval($options['l'] ?? $options['limit'] ?? 0);
 $debug = isset($options['d']) || isset($options['debug']);
 $update = isset($options['u']) || isset($options['update']);
+$noDelete = isset($options['no-delete']);
 
 $movies = json_decode(file_get_contents($config['movies']), true);
 if ($limit > 0) {
     $movies = array_slice($movies, 0, $limit);
 }
 
-$config['loupe']->deleteAllDocuments();
+if (!$noDelete) {
+    $config['loupe']->deleteAllDocuments();
+}
 
 $startTime = microtime(true);
